refactor(main -> Person) : Person class를 main에서 분리
